Refactor merge scopes
